+++ database/seeds/DatabaseSeeder.php
@@ -9,8 +9,13 @@ class DatabaseSeeder extends Seeder
      *
      * @return void
      */
-    public function run()
-    {
+    public function run() {
         // $this->call(UsersTableSeeder::class);
+        DB::table('users')->insert([
+            'first_name'=> str_random(4),
+            'last_name'=> str_random(6),
+            'email'    => str_random(3).(string)rand(1000,9999).'@rit.edu',
+            'password' => bcrypt('p@ssword')
+        ]);
     }
 }
